fix: moved object get into new object function module
